+++ kadai3-1/nas/simpletype/main.go
@@ -4,34 +4,9 @@ import (
 	"context"
 	"fmt"
 	"time"
+
+	"github.com/gopherdojo/dojo7/kadai3-1/nas/simpletype/pkg/exercise"
 )
-
-// Exercise has questions and now question number
-type Exercise struct {
-	Questions      []string
-	NowQuestionNum int
-}
-
-// Next return Question stil has questions or not
-func (e *Exercise) Next() bool {
-	if e.NowQuestionNum >= len(e.Questions) {
-		e.NowQuestionNum++
-		return false
-	}
-	e.NowQuestionNum++
-	return true
-}
-
-// Get return question
-func (e *Exercise) Get() string {
-	if e.NowQuestionNum == 0 {
-		return ""
-	}
-	if e.NowQuestionNum > len(e.Questions) {
-		return ""
-	}
-	return e.Questions[e.NowQuestionNum-1]
-}
 
 func main() {
 	run()
@@ -42,7 +17,7 @@ func run() {
 	t := 10 * time.Second
 	ctx, cancel := context.WithTimeout(bc, t)
 	defer cancel()
-	e := &Exercise{
+	e := &exercise.Exercise{
 		Questions: []string{"apple", "banana", "cat", "dog", "egg", "fish"},
 	}
 	if !e.Next() {
